fix: 나의 사전 컴포넌트화
