+++ src/types/dictionary.ts
@@ -1,0 +1,9 @@
+export interface IdiomData {
+  id: number | string;
+  english: string;
+  korean: string;
+  example: string;
+  usage: string[];
+  variations: string[];
+  showDetails?: boolean;
+}
